Add dagger2-aware roboelectric junit tests
NOTE!   These tests do not yet run.   This check-in is mainly to get the testing
infrastructure in place and make use of Dagger2's dependency injection to
inject non-production, test-only objects into the code.

MainActivty is launched , but as it has  no widgets to test, no test is implemented
at this stage, so the test will fail   The next check-ins should have some actual tests.

(Yes, I know, it's generally not a good idea to check in broken tests.)

NOTE:  You'll need to clean and rebuild and run the tests to make a red-squiggly
line in Android Studio go away.  Dagger2 will have to generate code before
AS can find it and suppress the line.   The build would actually fail if we hadn't
suppressed the error in gradle with these lines:

tasks.withType(Test) {
  testLogging {
    exceptionFormat "full"
    events "started", "skipped", "passed", "failed"
    showStandardStreams true
  }
}